D9/Ex2/requirements.txt
kafka-python
faust-streaming
